This enforces the security policy on maniphest tasks when they are added to a "special" project, which is currently hard-coded.
See http://fab.wmflabs.org/T95

This works well and avoids patching phabricator.
The custom policy is configured by editing the
default policy on the corresponding project,
each task simply takes on the policy of the project.
This behavior only affects tasks on projects named
"Security" or "RT" - this is hard-coded in the
isSpecialProject method.

Change-Id: I1faf75c70862555b0fa073b5202e79ee7a2ce99b